+++ libs/i18n/locales/en/translation.json
@@ -85,6 +85,7 @@
   "Configurations": "Configurations",
   "System image": "System image",
   "Fleet name": "Fleet name",
+  "Sources ({{size}})": "Sources ({{size}})",
   "Applications": "Applications",
   "System services": "System services",
   "Multiple owners": "Multiple owners",
@@ -187,7 +188,6 @@
   "Label selector": "Label selector",
   "Associated devices": "Associated devices",
   "Managed by": "Managed by",
-  "Sources ({{size}})": "Sources ({{size}})",
   "Managed by the resource sync {{resourceSyncName}}": "Managed by the resource sync {{resourceSyncName}}",
   "Create a fleet": "Create a fleet",
   "Import fleets": "Import fleets",
