+++ libs/ui-components/src/utils/devices.ts
@@ -1,8 +1,14 @@
-import { Device, ObjectMeta } from '@flightctl/types';
+import { Device, DeviceSpec, ObjectMeta } from '@flightctl/types';
 import { DeviceAnnotation, FlightCtlLabel } from '../types/extraTypes';
 import { TFunction } from 'i18next';
 import { toAPILabel } from './labels';
 import { getMetadataAnnotation } from './api';
+
+export type SourceItem = {
+  type: 'git' | 'inline' | 'secret';
+  name: string;
+  displayText: string;
+};
 
 const deviceFleetRegExp = /^Fleet\/(?<fleetName>.*)$/;
 
@@ -40,4 +46,24 @@ const getUpdatedDevice = (device: Device, newLabels: FlightCtlLabel[]): Device =
   };
 };
 
-export { getDeviceFleet, getUpdatedDevice, getMissingFleetDetails };
+const getSourceItems = (specConfigs: DeviceSpec['config'] | undefined): SourceItem[] => {
+  return (specConfigs ?? [])
+    .map((config) => {
+      let sourceItem: SourceItem;
+      if ('gitRef' in config) {
+        sourceItem = { type: 'git', name: config.gitRef.repository, displayText: config.name };
+      } else if ('secretRef' in config) {
+        sourceItem = {
+          type: 'secret',
+          name: config.name,
+          displayText: `${config.secretRef.namespace || ''}/${config.secretRef.name || ''}`,
+        };
+      } else {
+        sourceItem = { type: 'inline', name: config.name, displayText: config.name };
+      }
+      return sourceItem;
+    })
+    .filter((repoName) => !!repoName);
+};
+
+export { getDeviceFleet, getUpdatedDevice, getMissingFleetDetails, getSourceItems };
